Before digging in to issue #208, move the grade book GUI code into its own package.
