+++ grader/src/main/java/edu/pdx/cs410J/grader/GradeBook.java
@@ -357,7 +357,7 @@ public class GradeBook {
     return this.studentsStream().filter(s -> ssn.equals(s.getSsn())).findAny();
   }
 
-  static class LetterGradeRanges implements Iterable<LetterGradeRanges.LetterGradeRange> {
+  public static class LetterGradeRanges implements Iterable<LetterGradeRanges.LetterGradeRange> {
     private final Map<LetterGrade, LetterGradeRange> ranges = new TreeMap<>();
 
     private LetterGradeRanges() {
@@ -435,7 +435,7 @@ public class GradeBook {
       return this.ranges.values().spliterator();
     }
 
-    static class LetterGradeRange {
+    public static class LetterGradeRange {
       private final LetterGrade letterGrade;
       private int maximum;
       private int minimum;
