+++ grader/src/main/java/edu/pdx/cs410J/grader/GraderTools.java
@@ -3,6 +3,7 @@ package edu.pdx.cs410J.grader;
 import ch.qos.logback.classic.Level;
 import ch.qos.logback.classic.Logger;
 import com.google.common.annotations.VisibleForTesting;
+import edu.pdx.cs410J.grader.gradebook.GradeBookGUI;
 import edu.pdx.cs410J.grader.poa.ui.PlanOfAttackGrader;
 import org.slf4j.LoggerFactory;
